requirements.txt
supabase
python-dotenv
openai
ensta==5.2.5
moviepy==1.0.3